+++ instapy/instapy.py
@@ -1310,11 +1310,11 @@ class InstaPy:
         self.min_posts = min_posts if enabled is True else None
         self.max_posts = max_posts if enabled is True else None
 
-    def validate_user_call(self, user_name: str):
+    def validate_user_call(self, target_username: str):
         """ Short call of validate_username() function """
         validation, details = validate_username(
             self.browser,
-            user_name,
+            target_username,
             self.username,
             self.ignore_users,
             self.blacklist,
@@ -1492,7 +1492,13 @@ class InstaPy:
                 self.logger.info(link)
 
                 try:
-                    inappropriate, user_name, is_video, reason, scope = check_link(
+                    (
+                        inappropriate,
+                        target_username,
+                        is_video,
+                        reason,
+                        scope,
+                    ) = check_link(
                         self.browser,
                         link,
                         self.dont_like,
@@ -1512,7 +1518,7 @@ class InstaPy:
 
                     if not inappropriate and self.liking_approved:
                         # validate user
-                        validation, details = self.validate_user_call(user_name)
+                        validation, details = self.validate_user_call(target_username)
 
                         if validation is not True:
                             self.logger.info("--> Not a valid user: {}".format(details))
@@ -1524,7 +1530,7 @@ class InstaPy:
                         # try to like
                         like_state, msg = like_image(
                             self.browser,
-                            user_name,
+                            target_username,
                             self.blacklist,
                             self.logger,
                             self.logfolder,
@@ -1560,7 +1566,7 @@ class InstaPy:
                             # comments
                             if (
                                 self.do_comment
-                                and user_name not in self.dont_include
+                                and target_username not in self.dont_include
                                 and checked_img
                                 and commenting
                             ):
@@ -1569,6 +1575,7 @@ class InstaPy:
                                     if is_video
                                     else self.photo_comments
                                 )
+                                self.target_username = target_username
                                 success = process_comments(
                                     comments,
                                     temp_comments,
@@ -1577,6 +1584,7 @@ class InstaPy:
                                     self.min_comments,
                                     self.comments_mandatory_words,
                                     self.username,
+                                    self.target_username,
                                     self.blacklist,
                                     self.browser,
                                     self.logger,
@@ -1592,11 +1600,14 @@ class InstaPy:
                             # following
                             if (
                                 self.do_follow
-                                and user_name not in self.dont_include
+                                and target_username not in self.dont_include
                                 and checked_img
                                 and following
                                 and not follow_restriction(
-                                    "read", user_name, self.follow_times, self.logger
+                                    "read",
+                                    target_username,
+                                    self.follow_times,
+                                    self.logger,
                                 )
                             ):
 
@@ -1604,7 +1615,7 @@ class InstaPy:
                                     self.browser,
                                     "post",
                                     self.username,
-                                    user_name,
+                                    target_username,
                                     None,
                                     self.blacklist,
                                     self.logger,
@@ -1708,7 +1719,13 @@ class InstaPy:
                 self.logger.info(link)
 
                 try:
-                    inappropriate, user_name, is_video, reason, scope = check_link(
+                    (
+                        inappropriate,
+                        target_username,
+                        is_video,
+                        reason,
+                        scope,
+                    ) = check_link(
                         self.browser,
                         link,
                         self.dont_like,
@@ -1722,7 +1739,7 @@ class InstaPy:
                     )
                     if not inappropriate:
                         # validate user
-                        validation, details = self.validate_user_call(user_name)
+                        validation, details = self.validate_user_call(target_username)
                         if validation is not True:
                             self.logger.info(details)
                             not_valid_users += 1
@@ -1761,7 +1778,7 @@ class InstaPy:
 
                         if (
                             self.do_comment
-                            and user_name not in self.dont_include
+                            and target_username not in self.dont_include
                             and checked_img
                         ):
                             comments = self.comments + (
@@ -1769,6 +1786,7 @@ class InstaPy:
                             )
 
                             if comments:
+                                self.target_username = target_username
                                 success = process_comments(
                                     comments,
                                     temp_comments,
@@ -1777,6 +1795,7 @@ class InstaPy:
                                     self.min_comments,
                                     self.comments_mandatory_words,
                                     self.username,
+                                    self.target_username,
                                     self.blacklist,
                                     self.browser,
                                     self.logger,
@@ -1792,12 +1811,12 @@ class InstaPy:
                                     # try to follow
                                     if (
                                         self.do_follow
-                                        and user_name not in self.dont_include
+                                        and target_username not in self.dont_include
                                         and checked_img
                                         and following
                                         and not follow_restriction(
                                             "read",
-                                            user_name,
+                                            target_username,
                                             self.follow_times,
                                             self.logger,
                                         )
@@ -1806,7 +1825,7 @@ class InstaPy:
                                             self.browser,
                                             "post",
                                             self.username,
-                                            user_name,
+                                            target_username,
                                             None,
                                             self.blacklist,
                                             self.logger,
@@ -1928,7 +1947,13 @@ class InstaPy:
                 self.logger.info(link)
 
                 try:
-                    inappropriate, user_name, is_video, reason, scope = check_link(
+                    (
+                        inappropriate,
+                        target_username,
+                        is_video,
+                        reason,
+                        scope,
+                    ) = check_link(
                         self.browser,
                         link,
                         self.dont_like,
@@ -1948,7 +1973,7 @@ class InstaPy:
 
                     if not inappropriate and self.liking_approved:
                         # validate user
-                        validation, details = self.validate_user_call(user_name)
+                        validation, details = self.validate_user_call(target_username)
                         if validation is not True:
                             self.logger.info(details)
                             not_valid_users += 1
@@ -1959,7 +1984,7 @@ class InstaPy:
                         # try to like
                         like_state, msg = like_image(
                             self.browser,
-                            user_name,
+                            target_username,
                             self.blacklist,
                             self.logger,
                             self.logfolder,
@@ -1995,7 +2020,7 @@ class InstaPy:
                             # comments
                             if (
                                 self.do_comment
-                                and user_name not in self.dont_include
+                                and target_username not in self.dont_include
                                 and checked_img
                                 and commenting
                             ):
@@ -2004,6 +2029,7 @@ class InstaPy:
                                     if is_video
                                     else self.photo_comments
                                 )
+                                self.target_username = target_username
                                 success = process_comments(
                                     comments,
                                     temp_comments,
@@ -2012,6 +2038,7 @@ class InstaPy:
                                     self.min_comments,
                                     self.comments_mandatory_words,
                                     self.username,
+                                    self.target_username,
                                     self.blacklist,
                                     self.browser,
                                     self.logger,
@@ -2027,11 +2054,14 @@ class InstaPy:
                             # following
                             if (
                                 self.do_follow
-                                and user_name not in self.dont_include
+                                and target_username not in self.dont_include
                                 and checked_img
                                 and following
                                 and not follow_restriction(
-                                    "read", user_name, self.follow_times, self.logger
+                                    "read",
+                                    target_username,
+                                    self.follow_times,
+                                    self.logger,
                                 )
                             ):
 
@@ -2039,7 +2069,7 @@ class InstaPy:
                                     self.browser,
                                     "post",
                                     self.username,
-                                    user_name,
+                                    target_username,
                                     None,
                                     self.blacklist,
                                     self.logger,
@@ -2055,14 +2085,14 @@ class InstaPy:
                             if interact:
                                 self.logger.info(
                                     "--> User gonna be interacted: '{}'".format(
-                                        user_name
+                                        target_username
                                     )
                                 )
 
                                 # disable revalidating user in like_by_users
                                 with self.feature_in_feature("like_by_users", False):
                                     self.like_by_users(
-                                        user_name,
+                                        target_username,
                                         self.user_interact_amount,
                                         self.user_interact_random,
                                         self.user_interact_media,
@@ -2229,7 +2259,13 @@ class InstaPy:
                 self.logger.info(link)
 
                 try:
-                    inappropriate, user_name, is_video, reason, scope = check_link(
+                    (
+                        inappropriate,
+                        target_username,
+                        is_video,
+                        reason,
+                        scope,
+                    ) = check_link(
                         self.browser,
                         link,
                         self.dont_like,
@@ -2250,7 +2286,7 @@ class InstaPy:
                     if not inappropriate and self.liking_approved:
                         like_state, msg = like_image(
                             self.browser,
-                            user_name,
+                            target_username,
                             self.blacklist,
                             self.logger,
                             self.logfolder,
@@ -2284,7 +2320,7 @@ class InstaPy:
 
                             if (
                                 self.do_comment
-                                and user_name not in self.dont_include
+                                and target_username not in self.dont_include
                                 and checked_img
                                 and commenting
                             ):
@@ -2293,6 +2329,7 @@ class InstaPy:
                                     if is_video
                                     else self.photo_comments
                                 )
+                                self.target_username = target_username
                                 success = process_comments(
                                     comments,
                                     temp_comments,
@@ -2301,6 +2338,7 @@ class InstaPy:
                                     self.min_comments,
                                     self.comments_mandatory_words,
                                     self.username,
+                                    self.target_username,
                                     self.blacklist,
                                     self.browser,
                                     self.logger,
@@ -2503,7 +2541,13 @@ class InstaPy:
                 self.logger.info(link)
 
                 try:
-                    inappropriate, user_name, is_video, reason, scope = check_link(
+                    (
+                        inappropriate,
+                        target_username,
+                        is_video,
+                        reason,
+                        scope,
+                    ) = check_link(
                         self.browser,
                         link,
                         self.dont_like,
@@ -2543,7 +2587,7 @@ class InstaPy:
                         if self.do_like and liking and self.liking_approved:
                             like_state, msg = like_image(
                                 self.browser,
-                                user_name,
+                                target_username,
                                 self.blacklist,
                                 self.logger,
                                 self.logfolder,
@@ -2578,6 +2622,7 @@ class InstaPy:
                                         if is_video
                                         else self.photo_comments
                                     )
+                                    self.target_username = target_username
                                     success = process_comments(
                                         comments,
                                         temp_comments,
@@ -2586,6 +2631,7 @@ class InstaPy:
                                         self.min_comments,
                                         self.comments_mandatory_words,
                                         self.username,
+                                        self.target_username,
                                         self.blacklist,
                                         self.browser,
                                         self.logger,
@@ -2820,7 +2866,13 @@ class InstaPy:
                 self.logger.info(link)
 
                 try:
-                    inappropriate, user_name, is_video, reason, scope = check_link(
+                    (
+                        inappropriate,
+                        target_username,
+                        is_video,
+                        reason,
+                        scope,
+                    ) = check_link(
                         self.browser,
                         link,
                         self.dont_like,
@@ -2855,7 +2907,7 @@ class InstaPy:
                         if self.do_like and liking and self.liking_approved:
                             like_state, msg = like_image(
                                 self.browser,
-                                user_name,
+                                target_username,
                                 self.blacklist,
                                 self.logger,
                                 self.logfolder,
@@ -2889,6 +2941,7 @@ class InstaPy:
                                         if is_video
                                         else self.photo_comments
                                     )
+                                    self.target_username = target_username
                                     success = process_comments(
                                         comments,
                                         temp_comments,
@@ -2897,6 +2950,7 @@ class InstaPy:
                                         self.min_comments,
                                         self.comments_mandatory_words,
                                         self.username,
+                                        self.target_username,
                                         self.blacklist,
                                         self.browser,
                                         self.logger,
@@ -3973,7 +4027,7 @@ class InstaPy:
                         try:
                             (
                                 inappropriate,
-                                user_name,
+                                target_username,
                                 is_video,
                                 reason,
                                 scope,
@@ -3999,7 +4053,9 @@ class InstaPy:
                                 )
                             if not inappropriate and self.liking_approved:
                                 # validate user
-                                validation, details = self.validate_user_call(user_name)
+                                validation, details = self.validate_user_call(
+                                    target_username
+                                )
                                 if validation is not True:
                                     self.logger.info(details)
                                     not_valid_users += 1
@@ -4010,7 +4066,7 @@ class InstaPy:
                                 # try to like
                                 like_state, msg = like_image(
                                     self.browser,
-                                    user_name,
+                                    target_username,
                                     self.blacklist,
                                     self.logger,
                                     self.logfolder,
@@ -4050,7 +4106,7 @@ class InstaPy:
                                     # commenting
                                     if (
                                         self.do_comment
-                                        and user_name not in self.dont_include
+                                        and target_username not in self.dont_include
                                         and checked_img
                                         and commenting
                                     ):
@@ -4059,6 +4115,7 @@ class InstaPy:
                                             if is_video
                                             else self.photo_comments
                                         )
+                                        self.target_username = target_username
                                         success = process_comments(
                                             comments,
                                             temp_comments,
@@ -4067,6 +4124,7 @@ class InstaPy:
                                             self.min_comments,
                                             self.comments_mandatory_words,
                                             self.username,
+                                            self.target_username,
                                             self.blacklist,
                                             self.browser,
                                             self.logger,
@@ -4082,12 +4140,12 @@ class InstaPy:
                                     # following
                                     if (
                                         self.do_follow
-                                        and user_name not in self.dont_include
+                                        and target_username not in self.dont_include
                                         and checked_img
                                         and following
                                         and not follow_restriction(
                                             "read",
-                                            user_name,
+                                            target_username,
                                             self.follow_times,
                                             self.logger,
                                         )
@@ -4096,7 +4154,7 @@ class InstaPy:
                                             self.browser,
                                             "post",
                                             self.username,
-                                            user_name,
+                                            target_username,
                                             None,
                                             self.blacklist,
                                             self.logger,
@@ -4112,11 +4170,11 @@ class InstaPy:
                                     if interact:
                                         self.logger.info(
                                             "--> User gonna be interacted: "
-                                            "'{}'".format(user_name)
+                                            "'{}'".format(target_username)
                                         )
 
                                         self.like_by_users(
-                                            user_name,
+                                            target_username,
                                             self.user_interact_amount,
                                             self.user_interact_random,
                                             self.user_interact_media,
@@ -4149,7 +4207,7 @@ class InstaPy:
                                     self.logger.warning(
                                         "--> Ongoing Unfollow {}: unfollowing"
                                         " '{}' due to inappropriate content...".format(
-                                            inap_unfollow + 1, user_name
+                                            inap_unfollow + 1, target_username
                                         )
                                     )
 
@@ -4157,7 +4215,7 @@ class InstaPy:
                                         self.browser,
                                         "post",
                                         self.username,
-                                        user_name,
+                                        target_username,
                                         None,
                                         None,
                                         self.relationship_data,
@@ -4508,7 +4566,13 @@ class InstaPy:
                 self.logger.info(link)
 
                 try:
-                    inappropriate, user_name, is_video, reason, scope = check_link(
+                    (
+                        inappropriate,
+                        target_username,
+                        is_video,
+                        reason,
+                        scope,
+                    ) = check_link(
                         self.browser,
                         link,
                         self.dont_like,
@@ -4523,7 +4587,7 @@ class InstaPy:
 
                     if not inappropriate:
                         # validate user
-                        validation, details = self.validate_user_call(user_name)
+                        validation, details = self.validate_user_call(target_username)
                         if validation is not True:
                             self.logger.info(details)
                             not_valid_users += 1
@@ -4536,7 +4600,7 @@ class InstaPy:
                             self.browser,
                             "post",
                             self.username,
-                            user_name,
+                            target_username,
                             None,
                             self.blacklist,
                             self.logger,
@@ -4640,7 +4704,13 @@ class InstaPy:
                 self.logger.info(link)
 
                 try:
-                    inappropriate, user_name, is_video, reason, scope = check_link(
+                    (
+                        inappropriate,
+                        target_username,
+                        is_video,
+                        reason,
+                        scope,
+                    ) = check_link(
                         self.browser,
                         link,
                         self.dont_like,
@@ -4655,7 +4725,7 @@ class InstaPy:
 
                     if not inappropriate:
                         # validate user
-                        validation, details = self.validate_user_call(user_name)
+                        validation, details = self.validate_user_call(target_username)
                         if validation is not True:
                             self.logger.info(details)
                             not_valid_users += 1
@@ -4668,7 +4738,7 @@ class InstaPy:
                             self.browser,
                             "post",
                             self.username,
-                            user_name,
+                            target_username,
                             None,
                             self.blacklist,
                             self.logger,
@@ -4692,7 +4762,7 @@ class InstaPy:
                                     # disable following temporarily
                                     self.do_follow = False
                                     self.interact_by_users(
-                                        user_name,
+                                        target_username,
                                         self.user_interact_amount,
                                         self.user_interact_random,
                                         self.user_interact_media,
@@ -4764,7 +4834,7 @@ class InstaPy:
             self.logger.info("--> {}".format(url.encode("utf-8")))
 
             try:
-                inappropriate, user_name, is_video, reason, scope = check_link(
+                inappropriate, target_username, is_video, reason, scope = check_link(
                     self.browser,
                     url,
                     self.dont_like,
@@ -4784,7 +4854,7 @@ class InstaPy:
 
                 if not inappropriate and self.liking_approved:
                     # validate user
-                    validation, details = self.validate_user_call(user_name)
+                    validation, details = self.validate_user_call(target_username)
                     if validation is not True:
                         self.logger.info(details)
                         not_valid_users += 1
@@ -4795,7 +4865,7 @@ class InstaPy:
                     # try to like
                     like_state, msg = like_image(
                         self.browser,
-                        user_name,
+                        target_username,
                         self.blacklist,
                         self.logger,
                         self.logfolder,
@@ -4826,13 +4896,14 @@ class InstaPy:
 
                         if (
                             self.do_comment
-                            and user_name not in self.dont_include
+                            and target_username not in self.dont_include
                             and checked_img
                             and commenting
                         ):
                             comments = self.comments + (
                                 self.video_comments if is_video else self.photo_comments
                             )
+                            self.target_username = target_username
                             success = process_comments(
                                 comments,
                                 temp_comments,
@@ -4841,6 +4912,7 @@ class InstaPy:
                                 self.min_comments,
                                 self.comments_mandatory_words,
                                 self.username,
+                                self.target_username,
                                 self.blacklist,
                                 self.browser,
                                 self.logger,
@@ -4855,18 +4927,18 @@ class InstaPy:
 
                         if (
                             self.do_follow
-                            and user_name not in self.dont_include
+                            and target_username not in self.dont_include
                             and checked_img
                             and following
                             and not follow_restriction(
-                                "read", user_name, self.follow_times, self.logger
+                                "read", target_username, self.follow_times, self.logger
                             )
                         ):
                             follow_state, msg = follow_user(
                                 self.browser,
                                 "post",
                                 self.username,
-                                user_name,
+                                target_username,
                                 None,
                                 self.blacklist,
                                 self.logger,
@@ -4887,7 +4959,7 @@ class InstaPy:
                             # do interactions if any
                             if do_interact and self.user_interact_amount > 0:
                                 self.interact_by_users(
-                                    user_name,
+                                    target_username,
                                     self.user_interact_amount,
                                     self.user_interact_random,
                                     self.user_interact_media,
@@ -5317,7 +5389,7 @@ class InstaPy:
                     self.username, message, "post iteration", "info", self.logger
                 )
 
-                (inappropriate, user_name, is_video, reason, scope) = check_link(
+                (inappropriate, target_username, is_video, reason, scope) = check_link(
                     self.browser,
                     link,
                     self.dont_like,
@@ -5360,7 +5432,7 @@ class InstaPy:
                 # like the post before interacting on comments
                 image_like_state, msg = like_image(
                     self.browser,
-                    user_name,
+                    target_username,
                     self.blacklist,
                     self.logger,
                     self.logfolder,
@@ -5748,7 +5820,7 @@ class InstaPy:
                 post_link = "https://www.instagram.com/p/{}".format(pod_post_id)
                 web_address_navigator(self.browser, post_link)
 
-                inappropriate, user_name, is_video, reason, scope = check_link(
+                inappropriate, target_username, is_video, reason, scope = check_link(
                     self.browser,
                     post_link,
                     self.dont_like,
@@ -5761,21 +5833,21 @@ class InstaPy:
                     self.logger,
                 )
 
-                if user_name != self.username:
+                if target_username != self.username:
                     follow_state, msg = follow_user(
                         self.browser,
                         "post",
                         self.username,
-                        user_name,
+                        target_username,
                         None,
                         self.blacklist,
                         self.logger,
                         self.logfolder,
                     )
 
-                    self.dont_include.add(user_name)
+                    self.dont_include.add(target_username)
 
-                if not inappropriate and user_name != self.username:
+                if not inappropriate and target_username != self.username:
                     pods_like_percent = max(90, min(100, self.like_percentage))
                     liking = random.randint(0, 100) <= pods_like_percent
                     commenting = (
@@ -5785,7 +5857,7 @@ class InstaPy:
                     if liking:
                         like_state, msg = like_image(
                             self.browser,
-                            user_name,
+                            target_username,
                             self.blacklist,
                             self.logger,
                             self.logfolder,
@@ -5807,7 +5879,7 @@ class InstaPy:
 
                         comment_state, msg = comment_image(
                             self.browser,
-                            user_name,
+                            target_username,
                             comments,
                             self.blacklist,
                             self.logger,